Fixes test
